Added fixes and new features (QoL)
Features:
You can now automatically choose an option instead of having to type "a" or "a -i" for everything account. Edit auto_choice in the settings.json.

You can now disable the option to overwrite files. No more downloading the same file twice. Downside is that if there's a corrupt file, it will still be ignored. Set overwrite_files to false if you want this feature.

Fixes:
Increased the chance of grabbing the correct date. If the date is invalid in the json, the script will then try to grab the date from the URL. This only works when downloading videos.

When downloading large amount of files, some users may get a timeout. I made sure to create an infinite download loop so that the file will almost certaintly download. Downside is that the script may get stuck in the loop if the file actually doesn't exist. Any downloads that doesn't work will be sent to the errors.log.

Improvements:
You will notice the script will run fast after entering a user name. One of OnlyFan's API endpoints was slow, so I found a different way to grab the user ID.