Faster Scraping
I'm rewriting most of the script to make it look better. Half the time, I don't even know what I'm looking for.

Anyways, the script should be up to 100% faster since we'll only call the api once instead of three times and then search for image, video and audio files.