OFRenamer fix and increased speed
Basically the script was moving files to the new location, but deleting it straight after.

Some cloud services would get confused and start putting the remote file in the remote root folder because we deleted the local file before the cloud service could update the location.

This was a bug I had for a year now and I finally found the issue and fixed it and now I have to delete thousands of duplicate images and folders whilst lagging :)

You'll only notice the speed increase if scrape_names is False and scrape_paid_content is True. We skip getting the subscriber list if you're only scraping paid content.

Metadata kept duplicated posts that were deleted by the model. I've fixed that. Some metadata files should decrease in size.