Fixes: OFRenamer
Script no longer downloads files that are already downloaded unless overwritten is set to true.

Fixed renaming.

OnlyFans decided to let creators send posts to messages which would cause errors such as the FileNotFoundError. I added the "linked" column in the database. Any file that is linked will have "linked" in it's filename.