Changed create_signed_headers and json_requests
I removed all of the variables that wasn't needed in the create_signed_headers function so it's easier to read and copy.

json_requests is now under the session_manager class. I got rid of the majority of global variables.
I did this so we don't have to keep adding a session variable to a param.