Dynamic Rules
Script will grab signed_header rules from github so the end user won't have to keep updating the script.

OF's goal is to create download fatigue on the end user so they won't use scripts to archive data.

Right now the script uses inefficient logic to assign the rules. I'll update the script later, but I just wanted to put this out now.

Logic being used right now, is that on every request to a OnlyFans' api the script sends a request to Github to grab the rules, and then proceed with the request.