Fixes fallback to source link
I noticed that I can't download certain files when setting quality to 720p. Turns out that for files which don't have a 720p URL, the fallback source URL gets overwritten. Since there's quite a bit of logic to fetch the source URL independent of the quality setting, I assume this is a bug.