Implemented Asyncio
Script now uses async functions to handle network requests and writing to disk. I did this because it's faster.

Version of the script is still somewhat unstable when it comes to network requests, but I've scraped hundred of models with it and fixed multiple bugs.

max_threads in the config now only controls how many network requests to make at a time, but the script will use all available threads to process data, so if you're running out of RAM, I suggest you lower max threads.

I'll continue optimizing the script and making it as fast as possible.