Make psutil dependency not required
The module was used only to obtain the free space on disk but the same behavior can be reproduced in pure Python. Making the module not required helps the portability of the code as it not requires building or downloading prebuilt versions of the psutil module which may not available for some Python versions or systems.